FW-651. Move libs to opencoap and aes/ccm to boards/common.
