+++ openapps/opencoap/cborencoder.c
