+++ openapps/opencoap/hkdf.c
