+++ openapps/opencoap/hmac.c
