+++ openapps/opencoap/sha-private.h
