+++ openstack/SConscript
@@ -33,13 +33,6 @@ sources_c = [
     os.path.join('cross-layers','openqueue.c'),
     os.path.join('cross-layers','openrandom.c'),
     os.path.join('cross-layers','packetfunctions.c'),
-    os.path.join('cross-layers','openaes.c'),
-    os.path.join('cross-layers','openccms.c'),
-    os.path.join('cross-layers','usha.c'),
-    os.path.join('cross-layers','hkdf.c'),
-    os.path.join('cross-layers','hmac.c'),
-    os.path.join('cross-layers','sha224-256.c'),
-    os.path.join('cross-layers','cborencoder.c'),
 ]
 sources_h = [
     'openstack.h',
@@ -69,10 +62,6 @@ sources_h = [
     os.path.join('cross-layers','openqueue.h'),
     os.path.join('cross-layers','openrandom.h'),
     os.path.join('cross-layers','packetfunctions.h'),
-    os.path.join('cross-layers','openaes.h'),
-    os.path.join('cross-layers','openccms.h'),
-    os.path.join('cross-layers','sha.h'),
-    os.path.join('cross-layers','cborencoder.h'),
     #=== apps
     os.path.join('#','openapps','userialbridge','userialbridge.h'),
 ]
