+++ projects/python/SConscript.env
@@ -58,6 +58,7 @@ buildEnv.Append(
         os.path.join('#','build','python_gcc','inc'),
         # bsp
         os.path.join('#','build','python_gcc','bsp','boards'),
+        os.path.join('#','build','python_gcc','bsp','boards','common'),
         os.path.join('#','build','python_gcc','bsp','boards','python'),
         # drivers
         os.path.join('#','build','python_gcc','drivers','common'),
